requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis for caching
redis==5.0.1
hiredis==2.2.3

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# HTTP requests
httpx==0.25.2
aiohttp==3.9.1

# Geolocation and mapping
geopy==2.4.0
shapely==2.0.2

# Validation and serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS middleware
fastapi-cors==0.0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production server
gunicorn==21.2.0

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.7
prometheus-client==0.19.0
psutil==5.9.6
sentry-sdk[fastapi]==1.38.0

# Web scraping dependencies
scrapy==2.11.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
requests-html==0.10.0

# Task queue and scheduling
celery[redis]==5.3.4
redis==5.0.1
celery-beat==2.6.0
flower==2.0.1
kombu==5.3.4

# API clients and HTTP
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0
tenacity==8.2.3
pybreaker==1.0.2
aioredis==2.0.1
backoff==2.2.1

# Data validation and parsing
marshmallow==3.20.1
jsonschema==4.19.2
xmltodict==0.13.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3